// src/interfaces/UnitEquivalence.ts
export interface UnitEquivalence {
  idunit_equivalence?: number;
  base_unit?: string;
  des_unit?: string;
  multiple_quantity?: number;
  description?: string;
  status?: string;
  start_validity?: Date;
  end_validity?: Date;
  date_create?: Date;
  user_created?: string;
  date_updated?: Date;
  user_updated?: string;
}